Remove unnecessary middlware classes. Use WebSocketMiddleware with IWebHandler as a generic constrain.
