+++ src/Lykke.AlgoStore.Api/Startup.cs
@@ -29,8 +29,8 @@ using Lykke.AlgoStore.Algo.Charting;
 using Lykke.AlgoStore.Service.Security.Client;
 using Lykke.Service.Security.Client.AutorestClient.Models;
 using Lykke.AlgoStore.Api.RealTimeStreaming;
+using Lykke.AlgoStore.Api.RealTimeStreaming.DataStreamers.WebSockets.Handlers;
 using Lykke.AlgoStore.Api.RealTimeStreaming.DataStreamers.WebSockets.Middleware;
-using Lykke.AlgoStore.Api.RealTimeStreaming.DataTypes;
 
 namespace Lykke.AlgoStore.Api
 {
@@ -157,10 +157,10 @@ namespace Lykke.AlgoStore.Api
 
             app.UseWebSockets(webSocketOptions);
 
-            app.Map("/live/dummy", (_app) => _app.UseMiddleware<DummyWebSocketsMiddleware>());
-            app.Map("/live/candles", (_app) => _app.UseMiddleware<CandlesWebSocketsMiddleware>());
-            app.Map("/live/trades", (_app) => _app.UseMiddleware<WebSocketMiddleware<TradeChartingUpdate>>());
-            app.Map("/live/functions", (_app) => _app.UseMiddleware<WebSocketMiddleware<FunctionChartingUpdate>>());
+            app.Map("/live/dummy", (_app) => _app.UseMiddleware<WebSocketMiddleware<DummyWebSocketHandler>>());
+            app.Map("/live/candles", (_app) => _app.UseMiddleware<WebSocketMiddleware<CandlesWebSocketHandler>>());
+            app.Map("/live/trades", (_app) => _app.UseMiddleware<WebSocketMiddleware<WebSocketHandlerBase<TradeChartingUpdate>>>());
+            app.Map("/live/functions", (_app) => _app.UseMiddleware<WebSocketMiddleware<WebSocketHandlerBase<FunctionChartingUpdate>>>());
         }
 
         private async Task StartApplication(ISecurityClient securityClient)
